# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Mechanic Shop API"
  description: "A comprehensive REST API for managing a mechanic shop with customers, vehicles, mechanics, service tickets, and inventory. Includes JWT authentication, rate limiting, and caching."
  version: "2.0.0"
  contact:
    name: "API Support"
    url: "https://github.com/KimiCubb/Mechanics-Shop-db"
host: "mechanics-shop-db.onrender.com"
basePath: ""
schemes:
  - "https"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
    description: "JWT Bearer token. Format: Bearer <token>"
paths:
...
  /service-tickets:
    post:
      tags:
        - "Service Tickets"
      summary: "Create a new service ticket"
      description: "Create a new service ticket for a vehicle"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/ServiceTicketCreate"
      responses:
        201:
          description: "Service ticket created successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              service_ticket:
                $ref: "#/definitions/ServiceTicketResponse"
        400:
          description: "Invalid input"
        404:
          description: "Vehicle not found"

    get:
      tags:
        - "Service Tickets"
      summary: "Get all service tickets"
      description: "Retrieve all service tickets with details"
      responses:
        200:
          description: "Service tickets retrieved successfully"
          schema:
            $ref: "#/definitions/ServiceTicketsListResponse"
  /service-tickets/{ticket_id}:
    get:
      tags:
        - "Service Tickets"
      summary: "Get a specific service ticket"
      description: "Retrieve details for a specific service ticket including mechanics and parts"
      parameters:
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Service ticket retrieved successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              service_ticket:
                $ref: "#/definitions/ServiceTicketResponse"
        404:
          description: "Service ticket not found"

    put:
      tags:
        - "Service Tickets"
      summary: "Update service ticket details"
      description: "Update ticket status, description, and total cost"
      parameters:
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          schema:
            $ref: "#/definitions/ServiceTicketUpdate"
      responses:
        200:
          description: "Service ticket updated successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              service_ticket:
                $ref: "#/definitions/ServiceTicketResponse"
        400:
          description: "Invalid input"
        404:
          description: "Service ticket not found"

    delete:
      tags:
        - "Service Tickets"
      summary: "Delete a service ticket"
      description: "Remove a service ticket from the system"
      parameters:
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Service ticket deleted successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
        404:
          description: "Service ticket not found"
  /service-tickets/{ticket_id}/assign-mechanic/{mechanic_id}:
    put:
      tags:
        - "Service Tickets"
      summary: "Assign a mechanic to a service ticket"
      description: "Add a mechanic to work on a service ticket"
      parameters:
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
        - in: "path"
          name: "mechanic_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Mechanic assigned successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              service_ticket:
                $ref: "#/definitions/ServiceTicketResponse"
        400:
          description: "Mechanic already assigned to this ticket"
        404:
          description: "Service ticket or mechanic not found"
  /service-tickets/{ticket_id}/remove-mechanic/{mechanic_id}:
    put:
      tags:
        - "Service Tickets"
      summary: "Remove a mechanic from a service ticket"
      description: "Remove a mechanic from working on a service ticket"
      parameters:
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
        - in: "path"
          name: "mechanic_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Mechanic removed successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              service_ticket:
                $ref: "#/definitions/ServiceTicketResponse"
        404:
          description: "Service ticket or mechanic not found"
  /service-tickets/{ticket_id}/add-part:
    post:
      tags:
        - "Service Tickets"
      summary: "Add a part to a service ticket"
      description: "Add a part to a service ticket with specified quantity"
      parameters:
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/AddPartToTicket"
      responses:
        200:
          description: "Part added to ticket successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              service_ticket:
                $ref: "#/definitions/ServiceTicketResponse"
        400:
          description: "Invalid input or part already added to ticket"
        404:
          description: "Ticket or part not found"
  /service-tickets/{ticket_id}/remove-part/{part_id}:
    delete:
      tags:
        - "Service Tickets"
      summary: "Remove a part from a service ticket"
      description: "Remove a part from a service ticket"
      parameters:
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
        - in: "path"
          name: "part_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Part removed from ticket successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
        404:
          description: "Ticket or part not found"
  /service-tickets/{ticket_id}/edit:
    put:
      tags:
        - "Service Tickets"
      summary: "Edit mechanics on a service ticket"
      description: "Add or remove multiple mechanics from a service ticket"
      parameters:
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          schema:
            $ref: "#/definitions/EditTicketMechanics"
      responses:
        200:
          description: "Service ticket updated successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              service_ticket:
                $ref: "#/definitions/ServiceTicketResponse"
        400:
          description: "Invalid input"
        404:
          description: "Service ticket or mechanic not found"
  /service-tickets/{ticket_id}/parts:
    get:
      tags:
        - "Service Tickets"
      summary: "Get parts on a service ticket"
      description: "Retrieve all parts assigned to a service ticket"
      parameters:
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Parts retrieved successfully"
          schema:
            $ref: "#/definitions/TicketPartsResponse"
        404:
          description: "Service ticket not found"
...
definitions:
...
  ServiceTicketResponse:
    type: "object"
    properties:
      service_ticket_id:
        type: "integer"
      vehicle_id:
        type: "integer"
      date_in:
        type: "string"
        format: "date-time"
      date_out:
        type: "string"
        format: "date-time"
      description:
        type: "string"
      status:
        type: "string"
      total_cost:
        type: "number"
        format: "float"
      mechanics:
        type: "array"
        items:
          type: "object"
      parts:
        type: "array"
        items:
          type: "object"
  ServiceTicketsListResponse:
    type: "object"
    properties:
      message:
        type: "string"
      count:
        type: "integer"
      service_tickets:
        type: "array"
        items:
          $ref: "#/definitions/ServiceTicketResponse"
  ServiceTicketCreate:
    type: "object"
    required:
      - vehicle_id
      - description
    properties:
      vehicle_id:
        type: "integer"
        example: 1
      description:
        type: "string"
        example: "Oil change and tire rotation"
      status:
        type: "string"
        default: "Open"
        enum: ["Open", "In Progress", "Completed"]
      total_cost:
        type: "number"
        format: "float"
        default: 0.0
        example: 150.00
  ServiceTicketUpdate:
    type: "object"
    properties:
      description:
        type: "string"
        example: "Oil change, tire rotation, and brake inspection"
      status:
        type: "string"
        enum: ["Open", "In Progress", "Completed"]
        example: "In Progress"
      total_cost:
        type: "number"
        format: "float"
        example: 200.00
      date_out:
        type: "string"
        format: "date-time"
        example: "2024-01-15T18:00:00"
  AddPartToTicket:
    type: "object"
    required:
      - part_id
    properties:
      part_id:
        type: "integer"
        example: 1
      quantity:
        type: "integer"
        default: 1
        minimum: 1
        example: 2
  EditTicketMechanics:
    type: "object"
    properties:
      add_mechanics:
        type: "array"
        items:
          type: "integer"
        description: "Mechanic IDs to add"
        example: [1, 2]
      remove_mechanics:
        type: "array"
        items:
          type: "integer"
        description: "Mechanic IDs to remove"
        example: [3]
  TicketPart:
    type: "object"
    properties:
      part_id:
        type: "integer"
      name:
        type: "string"
      price:
        type: "number"
        format: "float"
      quantity:
        type: "integer"
  TicketPartsResponse:
    type: "object"
    properties:
      message:
        type: "string"
      ticket_id:
        type: "integer"
      parts:
        type: "array"
        items:
          $ref: "#/definitions/TicketPart"
...